requirements.txt
pycryptodome==3.20.0
Pillow==10.2.0
